requirements.txt
langchain_community
transformers
sentence-transformers
langchain_groq
faiss-cpu 
langchain-core
langchain-community  
langchain-openai
flask
blobfile
protobuf
semantic-router
llama-index